+++ Source/OxyPlot.Wpf/Series/Series.cs
@@ -6,10 +6,11 @@
 //   Abstract base class for series.
 // </summary>
 // --------------------------------------------------------------------------------------------------------------------
-
 namespace OxyPlot.Wpf
 {
+    using System;
     using System.Collections;
+    using System.Collections.Specialized;
     using System.Windows;
     using System.Windows.Controls;
     using System.Windows.Media;
@@ -48,12 +49,25 @@ namespace OxyPlot.Wpf
             "TrackerKey", typeof(string), typeof(Series), new PropertyMetadata(null, AppearanceChanged));
 
         /// <summary>
+        /// The event listener used to subscribe to ItemSource.CollectionChanged events
+        /// </summary>
+        private readonly EventListener eventListener;
+
+        /// <summary>
         /// Initializes static members of the <see cref="Series" /> class.
         /// </summary>
         static Series()
         {
             VisibilityProperty.OverrideMetadata(typeof(Series), new PropertyMetadata(Visibility.Visible, AppearanceChanged));
             BackgroundProperty.OverrideMetadata(typeof(Series), new FrameworkPropertyMetadata(null, AppearanceChanged));
+        }
+
+        /// <summary>
+        /// Initializes a new instance of the <see cref="Series" /> class.
+        /// </summary>
+        protected Series()
+        {
+            this.eventListener = new EventListener(this.OnCollectionChanged);
         }
 
         /// <summary>
@@ -171,6 +185,7 @@ namespace OxyPlot.Wpf
         protected override void OnItemsSourceChanged(IEnumerable oldValue, IEnumerable newValue)
         {
             base.OnItemsSourceChanged(oldValue, newValue);
+            this.SubscribeToCollectionChanged(oldValue, newValue);
             this.OnDataChanged();
         }
 
@@ -200,6 +215,74 @@ namespace OxyPlot.Wpf
             s.IsVisible = this.Visibility == Visibility.Visible;
             s.Font = this.FontFamily.ToString();
             s.TextColor = this.Foreground.ToOxyColor();
+        }
+
+        /// <summary>
+        /// If the ItemsSource implements INotifyCollectionChanged update the visual when the collection changes.
+        /// </summary>
+        /// <param name="oldValue">The old ItemsSource</param>
+        /// <param name="newValue">The new ItemsSource</param>
+        private void SubscribeToCollectionChanged(IEnumerable oldValue, IEnumerable newValue)
+        {
+            var collection = oldValue as INotifyCollectionChanged;
+            if (collection != null)
+            {
+                CollectionChangedEventManager.RemoveListener(collection, this.eventListener);
+            }
+
+            collection = newValue as INotifyCollectionChanged;
+            if (collection != null)
+            {
+                CollectionChangedEventManager.AddListener(collection, this.eventListener);
+            }
+        }
+
+        /// <summary>
+        /// Invalidate the view when the collection changes
+        /// </summary>
+        /// <param name="sender">The sender</param>
+        /// <param name="notifyCollectionChangedEventArgs">The collection changed args</param>
+        private void OnCollectionChanged(object sender, NotifyCollectionChangedEventArgs notifyCollectionChangedEventArgs)
+        {
+            this.OnDataChanged();
+        }
+
+        /// <summary>
+        /// Listens to and forwards any collection changed events
+        /// </summary>
+        private class EventListener : IWeakEventListener
+        {
+            /// <summary>
+            /// The delegate to forward to
+            /// </summary>
+            private readonly EventHandler<NotifyCollectionChangedEventArgs> onCollectionChanged;
+
+            /// <summary>
+            /// Initializes a new instance of the <see cref="EventListener" /> class
+            /// </summary>
+            /// <param name="onCollectionChanged">The handler</param>
+            public EventListener(EventHandler<NotifyCollectionChangedEventArgs> onCollectionChanged)
+            {
+                this.onCollectionChanged = onCollectionChanged;
+            }
+
+            /// <summary>
+            /// Receive a weak event
+            /// </summary>
+            /// <param name="managerType">The manager type</param>
+            /// <param name="sender">The sender</param>
+            /// <param name="e">The arguments</param>
+            /// <returns>Whether the event was handled or not</returns>
+            public bool ReceiveWeakEvent(Type managerType, object sender, EventArgs e)
+            {
+                if (managerType == typeof(CollectionChangedEventManager))
+                {
+                    this.onCollectionChanged(sender, (NotifyCollectionChangedEventArgs)e);
+                    return true;
+                }
+
+                return false;
+            }
         }
     }
 }